Extract dataset inspection estimators
